requirements.txt
affine==2.3.1
alphashape==1.3.1
attrs==22.2.0
certifi==2022.12.7
click==8.0.4
click-log==0.4.0
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
Fiona==1.8.22
geopandas==0.9.0
importlib-metadata==4.8.3
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
munch==2.5.0
networkx==2.5.1
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
POT==0.8.0
pyparsing==3.0.9
pyproj==3.0.1
python-dateutil==2.8.2
pytz==2022.7
rasterio==1.2.10
Rtree==0.9.7
scikit-learn==0.24.2
scipy==1.5.4
Shapely==1.8.5.post1
six==1.16.0
sklearn==0.0.post1
snuggs==1.4.7
threadpoolctl==3.1.0
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
trimesh==3.17.1
typing_extensions==4.1.1
zipp==3.6.0
